// src/types/app/email.ts
export type MailLabelType = 'personal' | 'company' | 'important' | 'private'

export type MailFolderType = 'inbox' | 'sent' | 'draft' | 'starred' | 'spam' | 'trash'
// ...
export type MailMetaType = {
    spam: number
    inbox: number
    draft: number
}

export type MailType = {
    id: number
    message: string
    subject: string
    isRead: boolean
    to: MailToType[]
    cc: string[] | []
    isStarred: boolean
    bcc: string[] | []
    from: MailFromType
    time: Date | string
    replies: MailType[]
    hasNextMail?: boolean
    folder: MailFolderType
    labels: MailLabelType[]
    hasPreviousMail?: boolean
    attachments: MailAttachmentType[]
}

export type MailFoldersArrType = {
    icon: ReactElement
    name: MailFolderType
}
export type MailFoldersObjType = {
    [key: string]: any[]
}

export type MailStore = {
    mails: MailType[] | null
    selectedMails: number[]
    currentMail: null | MailType
    mailMeta: null | MailMetaType
    filter: {
        q: string
        label: string
        folder: string
    }
}

export type MailLabelColors = {
    personal: string
    company: string
    private: string
    important: string
}

export type MailSidebarType = {
    hidden: boolean
    store: MailStore
    lgAbove: boolean
    // dispatch: Dispatch<any>
    leftSidebarOpen: boolean
    leftSidebarWidth: number
    mailDetailsOpen: boolean
    toggleComposeOpen: () => void
    handleLeftSidebarToggle: () => void
    setMailDetailsOpen: (val: boolean) => void
    handleSelectAllMail: (val: boolean) => void
}